+++ src/tools/index.ts
@@ -1,5 +1,6 @@
 export * from './get-folders.js';
 export * from './create-folder.js';
+export * from './delete-folder.js';
 export * from './list-mailbox.js';
 export * from './fetch-summaries.js';
 export * from './fetch-message.js';